Move auth logic to `moss-git-hosting-provider`
