+++ .gitignore
@@ -53,3 +53,6 @@ dist/
 crates/moss-git/oauth.json
 crates/moss-git/.env
 crates/moss-git/test-repo-lab/
+.env
+crates/moss-git-hosting-provider/github_oauth.json
+crates/moss-git-hosting-provider/gitlab_oauth.json
