+++ crates/moss-git-hosting-provider/src/gitlab.rs
@@ -1,1 +1,2 @@
-
+pub mod auth;
+pub mod client;
