--- crates/moss-git/src/adapters.rs
@@ -1,1 +1,0 @@
-pub mod auth;
