--- crates/moss-git/src/adapters/auth/oauth.rs
@@ -1,4 +1,0 @@
-pub mod github;
-pub mod gitlab;
-pub use github::*;
-pub use gitlab::*;
