--- crates/moss-git/src/models.rs
@@ -1,1 +1,0 @@
-pub mod oauth;
