--- crates/moss-git/src/models/oauth.rs
@@ -1,2 +1,0 @@
-pub mod github;
-pub mod gitlab;
