--- crates/moss-git/src/ports.rs
@@ -1,6 +1,0 @@
-use anyhow::Result;
-use git2::RemoteCallbacks;
-
-pub trait AuthAgent {
-    fn generate_callback<'a>(&'a self, cb: &mut RemoteCallbacks<'a>) -> Result<()>;
-}
